feat(asdf): use personal deno until the changes are merged

+++ setup/version-manager.sh
@@ -25,25 +25,42 @@ require 'lib-zlib'
 
 add_setup 'setup_version_manager'
 
+deno_pr_merged() {
+  if curl -sSL 'https://api.github.com/repos/asdf-community/asdf-deno/pulls/29' | grep -q '"merged": true'; then
+    return 0
+  else
+    return 1
+  fi
+}
+
 setup_version_manager() {
-  setup_version_manager__plugins="deno fzf neovim nodejs python rust shellcheck zig"
+  setup_version_manager__plugins="fzf neovim nodejs python rust shellcheck zig"
   step "Setting up version manager plugins..."
   set +e
   for setup_version_manager__plugin in $setup_version_manager__plugins; do
     bash -c ". $HOME/.asdf/asdf.sh && asdf plugin add $setup_version_manager__plugin"
   done
+  if deno_pr_merged; then
+    bash -c ". $HOME/.asdf/asdf.sh && asdf plugin add deno"
+  else
+    bash -c ". $HOME/.asdf/asdf.sh && asdf plugin add deno https://github.com/spywhere/asdf-deno.git"
+    bash -c ". $HOME/.asdf/asdf.sh && asdf plugin update deno build-from-source"
+  fi
   set -e
-  cmd cd "$HOME"
-  step "Installing version manager plugins..."
-  # asdf v0.9 will now error if the plugin is missing from .tool-versions
-  #   Ref: https://github.com/asdf-vm/asdf/issues/574
-  # Possible solutions:
-  #   - A new install flag: https://github.com/asdf-vm/asdf/issues/968#issuecomment-991106501
-  setup_version_manager__plugins="fzf nodejs python neovim"
-  set +e
-  for setup_version_manager__plugin in $setup_version_manager__plugins; do
-    bash -c ". $HOME/.asdf/asdf.sh && asdf install $setup_version_manager__plugin"
-  done
-  set -e
+
+  if test -f "$HOME/.tool-versions"; then
+    cmd cd "$HOME"
+    step "Installing version manager plugins..."
+    # asdf v0.9 will now error if the plugin is missing from .tool-versions
+    #   Ref: https://github.com/asdf-vm/asdf/issues/574
+    # Possible solutions:
+    #   - A new install flag: https://github.com/asdf-vm/asdf/issues/968#issuecomment-991106501
+    setup_version_manager__plugins="fzf nodejs python neovim"
+    set +e
+    for setup_version_manager__plugin in $setup_version_manager__plugins; do
+      bash -c ". $HOME/.asdf/asdf.sh && asdf install $setup_version_manager__plugin"
+    done
+    set -e
+  fi
   cmd cd "$CURRENT_DIR"
 }
